+++ backend/user-vault/schema/schema.sql
@@ -68,7 +68,7 @@ CREATE TABLE `path` (
 CREATE TABLE `path_resource` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT 'primary key',
   `path_no` varchar(32) NOT NULL DEFAULT '' COMMENT 'path no',
-  `res_code` varchar(32) NOT NULL DEFAULT '' COMMENT 'resource code',
+  `res_code` varchar(128) NOT NULL DEFAULT '' COMMENT 'resource code',
   `create_time` timestamp NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'when the record is created',
   `create_by` varchar(255) NOT NULL DEFAULT '' COMMENT 'who created this record',
   `update_time` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'when the record is updated',
@@ -80,7 +80,7 @@ CREATE TABLE `path_resource` (
 
 CREATE TABLE `resource` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT 'primary key',
-  `code` varchar(32) NOT NULL DEFAULT '' COMMENT 'resource code',
+  `code` varchar(128) NOT NULL DEFAULT '' COMMENT 'resource code',
   `name` varchar(32) NOT NULL DEFAULT '' COMMENT 'resource name',
   `create_time` timestamp NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'when the record is created',
   `create_by` varchar(255) NOT NULL DEFAULT '' COMMENT 'who created this record',
@@ -107,7 +107,7 @@ CREATE TABLE `role` (
 CREATE TABLE `role_resource` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT 'primary key',
   `role_no` varchar(32) NOT NULL DEFAULT '' COMMENT 'role no',
-  `res_code` varchar(32) NOT NULL DEFAULT '' COMMENT 'resource code',
+  `res_code` varchar(128) NOT NULL DEFAULT '' COMMENT 'resource code',
   `create_time` timestamp NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'when the record is created',
   `create_by` varchar(255) NOT NULL DEFAULT '' COMMENT 'who created this record',
   `update_time` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'when the record is updated',