+++ changes/changes.md
@@ -4,4 +4,4 @@ This change log only contains major changes to database schema.
 
 ## Release v0.0.3
 
-TODO
+1. [backend/user-vault/schema/v0.0.3.sql](../backend/user-vault/schema/v0.0.3.sql)